+++ src/main/java/com/supermartijn642/wirelesschargers/WirelessChargersClient.java
@@ -25,7 +25,7 @@ public class WirelessChargersClient {
             // Block entity renderers
             handler.registerCustomBlockEntityRenderer(type::getBlockEntityType, ChargerRenderer::new);
             // Baked item models
-            ResourceLocation location = new ModelResourceLocation(new ResourceLocation("wirelesschargers:" + type.getRegistryName()), "inventory");
+            ModelResourceLocation location = new ModelResourceLocation(ResourceLocation.fromNamespaceAndPath("wirelesschargers", type.getRegistryName()), "inventory");
             handler.registerModelOverwrite(location, CustomRendererBakedModelWrapper::wrap);
             // Item renderer
             //noinspection DataFlowIssue